+++ Lucc/Backend/x86_64Context.cpp
@@ -454,7 +454,7 @@ namespace lucc
 
 	std::string x86_64Context::DeclareString(char const* str)
 	{
-		static size_t string_unique_id = 0;
+		static uint64 string_unique_id = 0;
 		std::string str_label = std::format("L_str_{}", string_unique_id++);
 		Emit<Const>("{} byte \"{}\",0", str_label, str);
 		return str_label;
